DY - fix data attempt #Y - arrrgh

+++ data/setup/dev/cleandata.sql
@@ -33,13 +33,11 @@ DELETE FROM requisition_group_program_schedules;
 DELETE FROM requisition_group_members;
 DELETE FROM requisition_groups;
 
-DELETE FROM role_assignments;
+DELETE FROM role_assignments WHERE userId != 1;
 DELETE FROM user_password_reset_tokens;
 DELETE FROM users WHERE username != 'Admin123';
-DELETE FROM role_rights WHERE roleid= (SELECT id FROM roles WHERE name = 'Admin')
-   AND rightName IN ('MANAGE_GEOGRAPHIC_ZONE','MANAGE_SUPERVISORY_NODE','MANAGE_REQUISITION_GROUP','MANAGE_FACILITY_APPROVED_PRODUCT','MANAGE_PRODUCT');
-DELETE FROM role_rights WHERE roleid IN (SELECT id FROM roles WHERE name IN ('FacilityHead', 'ReportViewer', 'Supervisor')) ;
-DELETE FROM roles WHERE name IN ('FacilityHead', 'ReportViewer', 'Supervisor');
+DELETE FROM role_rights WHERE roleid != 1;
+DELETE FROM roles WHERE id != 1;
 
 DELETE FROM supervisory_nodes;
